+++ neoforge/src/main/java/crystalspider/harvestwithease/ModLoader.java
@@ -1,0 +1,33 @@
+package crystalspider.harvestwithease;
+
+import crystalspider.harvestwithease.config.ModConfig;
+import net.minecraft.resources.ResourceLocation;
+import net.neoforged.fml.ModLoadingContext;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig.Type;
+import net.neoforged.neoforge.network.NetworkRegistry.ChannelBuilder;
+import net.neoforged.neoforge.network.simple.SimpleChannel;
+
+/**
+ * Harvest with ease mod loader.
+ */
+@Mod(ModLoader.MOD_ID)
+public class ModLoader {
+  /**
+   * ID of this mod.
+   */
+  public static final String MOD_ID = "harvestwithease";
+
+  /**
+   * Network channel protocol version.
+   */
+  public static final String PROTOCOL_VERSION = "1.20.2-8.0";
+  /**
+   * {@link SimpleChannel} instance for compatibility client-server.
+   */
+  public static final SimpleChannel INSTANCE = ChannelBuilder.named(new ResourceLocation(MOD_ID, "main")).networkProtocolVersion(() -> PROTOCOL_VERSION).clientAcceptedVersions(PROTOCOL_VERSION::equals).serverAcceptedVersions(PROTOCOL_VERSION::equals).simpleChannel();
+
+  public ModLoader() {
+    ModLoadingContext.get().registerConfig(Type.COMMON, ModConfig.SPEC);
+  }
+}