+++ CHANGELOG.md
@@ -8,6 +8,26 @@ and this project adheres to [Forge Semantic Versioning](https://mcforge.readthed
 ## [Unreleased]
 - Nothing new.
 
+## [1.20.2-8.0.0.0] - 2023/11/24
+- Fixed [#28](https://github.com/Nyphet/harvest-with-ease/issues/28), Farmer's Delight tomatoes not giving xp when harvested.
+- Implemented [#30](https://github.com/Nyphet/harvest-with-ease/issues/30), multi-harvest.
+
+## [1.20.1-8.0.0.0] - 2023/11/24
+- Fixed [#28](https://github.com/Nyphet/harvest-with-ease/issues/28), Farmer's Delight tomatoes not giving xp when harvested.
+- Implemented [#30](https://github.com/Nyphet/harvest-with-ease/issues/30), multi-harvest.
+
+## [1.19.4-8.0.0.0] - 2023/11/24
+- Fixed [#28](https://github.com/Nyphet/harvest-with-ease/issues/28), Farmer's Delight tomatoes not giving xp when harvested.
+- Implemented [#30](https://github.com/Nyphet/harvest-with-ease/issues/30), multi-harvest.
+
+## [1.19.2-8.0.0.0] - 2023/11/24
+- Fixed [#28](https://github.com/Nyphet/harvest-with-ease/issues/28), Farmer's Delight tomatoes not giving xp when harvested.
+- Implemented [#30](https://github.com/Nyphet/harvest-with-ease/issues/30), multi-harvest.
+
+## [1.18.2-8.0.0.0] - 2023/11/24
+- Fixed [#28](https://github.com/Nyphet/harvest-with-ease/issues/28), Farmer's Delight tomatoes not giving xp when harvested.
+- Implemented [#30](https://github.com/Nyphet/harvest-with-ease/issues/30), multi-harvest.
+
 ## [1.20.2-7.1.0.0] - 2023/10/31
 - Implemented [#27](https://github.com/Nyphet/harvest-with-ease/issues/27).
 
@@ -260,9 +280,11 @@ It must be an integer value and will be effective only if greater than 0 and req
 [Unreleased]: https://github.com/Nyphet/harvest-with-ease
 [README]: https://github.com/Nyphet/harvest-with-ease#readme
 
+[1.20.2-8.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.2-8.0.0.0
 [1.20.2-7.1.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.2-7.1.0.0
 [1.20.2-7.0.0.2]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.2-7.0.0.2
 
+[1.20.1-8.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.1-8.0.0.0
 [1.20.1-7.1.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.1-7.1.0.0
 [1.20.1-7.0.0.2]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.1-7.0.0.2
 [1.20.1-7.0.0.1]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20.1-7.0.0.1
@@ -272,6 +294,7 @@ It must be an integer value and will be effective only if greater than 0 and req
 [1.20-7.0.0.1]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20-7.0.0.1
 [1.20-7.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.20-7.0.0.0
 
+[1.19.4-8.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.4-8.0.0.0
 [1.19.4-6.1.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.4-6.1.0.0
 [1.19.4-6.0.1.4]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.4-6.0.1.4
 [1.19.4-6.0.1.3]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.4-6.0.1.3
@@ -287,6 +310,7 @@ It must be an integer value and will be effective only if greater than 0 and req
 [1.19.3-5.0.0.1]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.3-5.0.0.1
 [1.19.3-5.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.3-5.0.0.0
 
+[1.19.2-8.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.2-8.0.0.0
 [1.19.2-6.1.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.2-6.1.0.0
 [1.19.2-6.0.1.4-final]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.2-6.0.1.4-final
 [1.19.2-6.0.1.3-final]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.2-6.0.1.3-final
@@ -304,12 +328,13 @@ It must be an integer value and will be effective only if greater than 0 and req
 [1.19.1-4.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.1-4.0.0.0
 [1.19.1-3.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19.1-3.0.0.0
 
-[1.19-4.0.0.1-final]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-4.0.0.1-final
-[1.19-4.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-4.0.0.0
+[1.19-2.0.0.1-final]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-2.0.0.1-final
+[1.19-2.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-2.0.0.0
 [1.19-3.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-3.0.0.0
 [1.19-2.0.0.1]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-2.0.0.1
 [1.19-2.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.19-2.0.0.0
 
+[1.18.2-8.0.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.18.2-8.0.0.0
 [1.18.2-6.1.0.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.18.2-6.1.0.0
 [1.18.2-6.0.1.3]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.18.2-6.0.1.3
 [1.18.2-6.0.1.2]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.18.2-6.0.1.2